--- a/ibis/pandas/aggcontext.py
+++ b/ibis/pandas/aggcontext.py
@@ -219,7 +219,6 @@ import functools
 import itertools
 import operator
 
-import numpy as np
 import pandas as pd
 from pandas.core.groupby import SeriesGroupBy
 
@@ -328,6 +327,83 @@ class Window(AggregationContext):
         )
         self.construct_window = operator.methodcaller(kind, *args, **kwargs)
 
+    def _agg_built_in(self, frame, windowed, function, *args, **kwargs):
+        assert isinstance(function, str)
+        method = operator.methodcaller(function, *args, **kwargs)
+
+        max_lookback = self.max_lookback
+        if max_lookback is not None:
+            agg_method = method
+
+            def sliced_agg(s):
+                return agg_method(s.iloc[-max_lookback:])
+
+            method = operator.methodcaller('apply', sliced_agg, raw=False)
+
+        result = method(windowed)
+        index = result.index
+        result.index = pd.MultiIndex.from_arrays(
+            [frame.index]
+            + list(map(index.get_level_values, range(index.nlevels))),
+            names=[frame.index.name] + index.names,
+        )
+        return result
+
+    def _agg_udf(self, grouped_data, windowed, function, *args, **kwargs):
+        # Use custom logic to computing rolling window UDF instead of
+        # using pandas's rolling function.
+        # This is because pandas's rolling function doesn't support
+        # multi param UDFs.
+
+        def create_input_gen(grouped_series, window_size):
+            # create a generator for each input series
+            # the generator will yield a slice of the
+            # input series for each valid window
+            data = getattr(grouped_series, 'obj', grouped_series).values
+            window_size_array = window_size.values
+            for i in range(len(window_size_array)):
+                k = window_size.index[i]
+                yield data[k - window_size_array[i] + 1 : k + 1]
+
+        obj = getattr(grouped_data, 'obj', grouped_data)
+
+        # Compute window indices and manually roll
+        # over the window.
+        # If an window has only nan values, we output nan for
+        # the window result. This follows pandas rolling apply
+        # behavior.
+        raw_window_size = windowed.apply(len, raw=True).reset_index(drop=True)
+        mask = ~(raw_window_size.isna())
+        window_size = raw_window_size[mask].astype('i8')
+        window_size_array = window_size.values
+
+        # If there is no args, then the UDF only takes a single
+        # input which is defined by grouped_data
+        # This is a complication due to the lack of standard
+        # way to pass multiple input Series/SeriesGroupBy
+        # to AggregationContext.agg()
+        inputs = args if len(args) > 0 else [grouped_data]
+
+        input_gens = list(
+            create_input_gen(arg, window_size)
+            if isinstance(arg, (pd.Series, SeriesGroupBy))
+            else itertools.repeat(arg)
+            for arg in inputs
+        )
+
+        valid_result = pd.Series(
+            function(*(next(gen) for gen in input_gens))
+            for i in range(len(window_size_array))
+        )
+
+        valid_result = pd.Series(valid_result)
+        valid_result.index = window_size.index
+        result = pd.Series(index=mask.index, dtype=self.dtype)
+        result[mask] = valid_result
+        result.index = obj.index
+
+        return result
+
     def agg(self, grouped_data, function, *args, **kwargs):
         # avoid a pandas warning about numpy arrays being passed through
         # directly
@@ -361,117 +437,46 @@ class Window(AggregationContext):
                     raw=True,
                 )
         else:
-            # do mostly the same thing as if we did NOT have a grouping key,
-            # but don't call the callable just yet. See below where we call it.
+            # get the DataFrame from which the operand originated
+            # (passed in when constructing this context object in
+            # execute_node(ops.WindowOp))
+            parent = self.parent
+            frame = getattr(parent, 'obj', parent)
+            obj = getattr(grouped_data, 'obj', grouped_data)
+            name = obj.name
+            if frame[name] is not obj:
+                name = "{}_{}".format(name, ibis.util.guid())
+                frame[name] = obj
+
+            # set the index to our order_by keys and append it to the existing
+            # index
+            # TODO: see if we can do this in the caller, when the context
+            # is constructed rather than pulling out the data
+            columns = group_by + order_by + [name]
+            indexed_by_ordering = frame.loc[:, columns].set_index(order_by)
+
+            # regroup if needed
+            if group_by:
+                grouped_frame = indexed_by_ordering.groupby(group_by)
+            else:
+                grouped_frame = indexed_by_ordering
+            grouped = grouped_frame[name]
+
+            # perform the per-group rolling operation
+            windowed = self.construct_window(grouped)
+
             if callable(function):
-                method = operator.methodcaller(
-                    'apply', make_applied_function(function, args, kwargs)
+                result = self._agg_udf(
+                    grouped_data, windowed, function, *args, **kwargs
                 )
             else:
-                assert isinstance(function, str)
-                method = operator.methodcaller(function, *args, **kwargs)
-
-            max_lookback = self.max_lookback
-            if max_lookback is not None:
-                agg_method = method
-
-                def sliced_agg(s):
-                    return agg_method(s.iloc[-max_lookback:])
-
-                method = operator.methodcaller('apply', sliced_agg, raw=False)
-
-        # get the DataFrame from which the operand originated
-        # (passed in when constructing this context object in
-        # execute_node(ops.WindowOp))
-        parent = self.parent
-        frame = getattr(parent, 'obj', parent)
-        obj = getattr(grouped_data, 'obj', grouped_data)
-        name = obj.name
-        if frame[name] is not obj:
-            name = "{}_{}".format(name, ibis.util.guid())
-            frame[name] = obj
-
-        # set the index to our order_by keys and append it to the existing
-        # index
-        # TODO: see if we can do this in the caller, when the context
-        # is constructed rather than pulling out the data
-        columns = group_by + order_by + [name]
-        indexed_by_ordering = frame.loc[:, columns].set_index(order_by)
-
-        # regroup if needed
-        if group_by:
-            grouped_frame = indexed_by_ordering.groupby(group_by)
-        else:
-            grouped_frame = indexed_by_ordering
-        grouped = grouped_frame[name]
-
-        # perform the per-group rolling operation
-        windowed = self.construct_window(grouped)
-
-        if callable(function):
-            # Use custom logic to computing rolling window UDF instead of
-            # using pandas's rolling function.
-            # This is because pandas's rolling function doesn't support
-            # multi param UDFs.
-            def create_input_gen(grouped_series, window_size):
-                # create a generator for each input series
-                # the generator will yield a slice of the
-                # input series for each valid window
-                data = getattr(grouped_series, 'obj', grouped_series).values
-                window_size_array = window_size.values
-                for i in range(len(window_size_array)):
-                    k = window_size.index[i]
-                    yield data[k - window_size_array[i] + 1 : k + 1]
-
-            # Compute window indices and manually roll
-            # over the window.
-            # If an window has only nan values, we output nan for
-            # the window result. This follows pandas rolling apply
-            # behavior.
-            raw_window_size = windowed.apply(len, raw=True).reset_index(
-                drop=True
-            )
-            mask = ~(raw_window_size.isna())
-            window_size = raw_window_size[mask].astype('i8')
-            window_size_array = window_size.values
-
-            # If there is no args, then the UDF only takes a single
-            # input which is defined by grouped_data
-            # This is a complication due to the lack of standard
-            # way to pass multiple input Series/SeriesGroupBy
-            # to AggregationContext.agg()
-            inputs = args if len(args) > 0 else [grouped_data]
-
-            input_gens = list(
-                create_input_gen(arg, window_size)
-                if isinstance(arg, (pd.Series, SeriesGroupBy))
-                else itertools.repeat(arg)
-                for arg in inputs
-            )
-
-            valid_result = pd.Series(
-                function(*(next(gen) for gen in input_gens))
-                for i in range(len(window_size_array))
-            )
-
-            valid_result = pd.Series(valid_result)
-            valid_result.index = window_size.index
-            result = pd.Series(np.repeat(None, len(obj)))
-            result[mask] = valid_result
-            result.index = obj.index
-        else:
-            result = method(windowed)
-            index = result.index
-            result.index = pd.MultiIndex.from_arrays(
-                [frame.index]
-                + list(map(index.get_level_values, range(index.nlevels))),
-                names=[frame.index.name] + index.names,
-            )
-
-        try:
-            return result.astype(self.dtype, copy=False)
-        except (TypeError, ValueError):
-            return result
+                result = self._agg_built_in(
+                    frame, windowed, function, *args, **kwargs
+                )
+            try:
+                return result.astype(self.dtype, copy=False)
+            except (TypeError, ValueError):
+                return result
 
 
 class Cumulative(Window):