… benchmarks/benchmarks.py
… docs/source/release.rst
… ibis/pandas/execution/tests/test_arrays.py
… ibis/pandas/udf.py