This PR addresses issue https://github.com/ibis-project/ibis/issues/1998

Currently, when using UDAF with rolling window, pandas backend will throw an exception:

```
import ibis
import pandas as pd
import numpy as np

from ibis.pandas.udf import udf
import ibis.expr.datatypes as dt

client = ibis.pandas.connect({'table': pd.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6], 'key': ['a', 'a', 'a']})})
t = client.table('table')
w = ibis.trailing_window(preceding=1, order_by='key', group_by='key')
#w = ibis.window(group_by='key')


@udf.reduction(input_type=[dt.double, dt.double], output_type=dt.double)
def my_average(v, w):
    return np.average(v, weights=w)

t = t.mutate(new_col=my_average(t.a, t.b).over(w))

t.execute()
```

With this PR, it will support this use case.

The main idea of this PR is that instead of using `groupby().rolling().apply(func)` to compute the result, we use `groupby().rolling().apply(len, raw=True)` to get the size of each window, and then manually apply `func` to each window in a Python for-loop. This way, we work around issue that `groupby().rolling().apply(func)` can only take function that apply on a single series.

Benchmarks
------------

Before the change:

```
In [4]: %time pandas.time_low_card_window_analytics_udf()                                                                                                     
CPU times: user 898 ms, sys: 248 ms, total: 1.15 s
Wall time: 1.15 s

In [5]: %time pandas.time_high_card_window_analytics_udf()                                                                                                    
CPU times: user 12.4 s, sys: 324 ms, total: 12.7 s
Wall time: 12.8 s

In [6]: %time pandas.time_low_card_grouped_rolling()                                                                                                          
CPU times: user 7.99 s, sys: 1.95 s, total: 9.94 s
Wall time: 10.1 s

In [7]: %time pandas.time_high_card_grouped_rolling()                                                                                                         
CPU times: user 15.3 s, sys: 1.83 s, total: 17.1 s
Wall time: 17.3 s

In [8]: %time pandas.time_low_card_grouped_rolling_udf()                                                                                                      
CPU times: user 2min 1s, sys: 1.96 s, total: 2min 3s
Wall time: 2min 4s

In [9]: %time pandas.time_high_card_grouped_rolling_udf()                                                                                                     
CPU times: user 1min 41s, sys: 2.17 s, total: 1min 43s
Wall time: 1min 43s
```

After the change:

```
time pandas.time_low_card_window_analytics_udf()
CPU times: user 856 ms, sys: 235 ms, total: 1.09 s
Wall time: 1.08 s
time pandas.time_high_card_window_analytics_udf()
CPU times: user 12.7 s, sys: 316 ms, total: 13 s
Wall time: 13.1 s
time pandas.time_low_card_grouped_rolling()
CPU times: user 9.3 s, sys: 3.15 s, total: 12.4 s
Wall time: 13 s
time pandas.time_high_card_grouped_rolling()
CPU times: user 15.5 s, sys: 1.9 s, total: 17.4 s
Wall time: 17.5 s
time pandas.time_low_card_grouped_rolling_udf()
CPU times: user 1min 26s, sys: 1.74 s, total: 1min 27s
Wall time: 1min 28s
time pandas.time_high_card_grouped_rolling_udf()
CPU times: user 1min 4s, sys: 1.79 s, total: 1min 6s
Wall time: 1min 6s
```