@icexelloss 2019-12-04
@jreback Haven't addressed all comments. Will ping again when it's done.